[Sri|Yin] Add suppression for vulnerabilities related to `jackson-databind` and `micrometer-tracing-bridge-brave`
Regarding the jackson-databind vulnerability, the vendor's perspective is that this is not a valid vulnerability report, because the steps of constructing a cyclic data structure and trying to serialize it cannot be achieved by an external attacker. As such, our suggestion is we could just suppress this.

As for the vulnerability caused by micrometer-tracing-bridge-brave it seems to be related to Brave Browser before 1.43.34 and there does not seem to be a patch for this on micrometer.